+++ providers/v1/gitlab/gitlab.go
@@ -19,25 +19,18 @@ package gitlab
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
 	"sort"
 	"strconv"
-	"strings"
 
-	"github.com/tidwall/gjson"
 	gitlab "gitlab.com/gitlab-org/api/client-go"
-	corev1 "k8s.io/api/core/v1"
-	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
 
 	esv1 "github.com/external-secrets/external-secrets/apis/externalsecrets/v1"
 	"github.com/external-secrets/external-secrets/runtime/constants"
-	"github.com/external-secrets/external-secrets/runtime/esutils"
 	"github.com/external-secrets/external-secrets/runtime/esutils/resolvers"
-	"github.com/external-secrets/external-secrets/runtime/find"
 	"github.com/external-secrets/external-secrets/runtime/metrics"
 )
 
@@ -91,7 +84,7 @@ func (a ProjectGroupPathSorter) Less(i, j int) bool { return len(a[i].FullPath) 
 
 var log = ctrl.Log.WithName("provider").WithName("gitlab")
 
-// Set gitlabBase credentials to Access Token.
+// getAuth retrieves the GitLab access token from the Kubernetes secret.
 func (g *gitlabBase) getAuth(ctx context.Context) (string, error) {
 	return resolvers.SecretKeyRef(
 		ctx,
@@ -101,838 +94,40 @@ func (g *gitlabBase) getAuth(ctx context.Context) (string, error) {
 		&g.store.Auth.SecretRef.AccessToken)
 }
 
-// PushSecret creates or updates a variable in GitLab.
-func (g *gitlabBase) PushSecret(ctx context.Context, secret *corev1.Secret, psd esv1.PushSecretData) error {
-	if esutils.IsNil(g.projectVariablesClient) && esutils.IsNil(g.groupVariablesClient) {
-		return errors.New(errUninitializedGitlabProvider)
+// getVariables retrieves a project variable with automatic wildcard retry on 404.
+func (g *gitlabBase) getVariables(ref esv1.ExternalSecretDataRemoteRef, vopts *gitlab.GetProjectVariableOptions) (*gitlab.ProjectVariable, error) {
+	// First attempt to get the variable
+	data, _, err := g.projectVariablesClient.GetVariable(g.store.ProjectID, ref.Key, vopts)
+	metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabProjectVariableGet, err)
+
+	// If successful, return immediately
+	if err == nil {
+		return data, nil
 	}
 
-	value, err := esutils.ExtractSecretData(psd, secret)
-	if err != nil {
-		return fmt.Errorf("failed to extract secret data: %w", err)
+	// If not a "not found" error or environment is already wildcard, return the error
+	if !errors.Is(err, gitlab.ErrNotFound) || isEmptyOrWildcard(g.store.Environment) {
+		return nil, err
 	}
 
-	// Get the remote key and replace hyphens with underscores for GitLab API
-	remoteKey := strings.ReplaceAll(psd.GetRemoteKey(), "-", "_")
-
-	// Extract metadata to check for groupID or projectID overrides
-	metadata := psd.GetMetadata()
-	groupID, projectID := extractTargetFromMetadata(metadata)
-
-	// Determine which type of variable to create/update
-	// Priority: metadata projectID > metadata groupID > store projectID > store groupIDs
-	if projectID != "" {
-		return g.pushProjectVariableWithID(projectID, remoteKey, string(value), psd)
-	}
-
-	if groupID != "" {
-		err := g.pushGroupVariable(groupID, remoteKey, string(value), psd)
-		if err != nil {
-			return fmt.Errorf("failed to push to group %q: %w", groupID, err)
-		}
-		return nil
-	}
-
-	// Fall back to store configuration
-	if g.store.ProjectID != "" {
-		return g.pushProjectVariable(remoteKey, string(value), psd)
-	}
-
-	// If no project ID is set, try to push to the first group
-	if len(g.store.GroupIDs) > 0 {
-		err := g.pushGroupVariable(g.store.GroupIDs[0], remoteKey, string(value), psd)
-		if err != nil {
-			return fmt.Errorf("failed to push to group %q: %w", g.store.GroupIDs[0], err)
-		}
-		return nil
-	}
-
-	return fmt.Errorf("%s for push operation", errNoProjectOrGroup)
-}
-
-// extractTargetFromMetadata extracts groupID and projectID from PushSecret metadata.
-// Returns empty strings if metadata is nil or doesn't contain the fields.
-func extractTargetFromMetadata(metadata *apiextensionsv1.JSON) (groupID, projectID string) {
-	if metadata == nil {
-		return "", ""
-	}
-
-	var metadataMap map[string]interface{}
-	if err := json.Unmarshal(metadata.Raw, &metadataMap); err != nil {
-		return "", ""
-	}
-
-	if gid, ok := metadataMap["groupID"].(string); ok {
-		groupID = gid
-	}
-	if pid, ok := metadataMap["projectID"].(string); ok {
-		projectID = pid
-	}
-
-	return groupID, projectID
-}
-
-// pushProjectVariableWithID creates or updates a project-level variable with a specific project ID.
-func (g *gitlabBase) pushProjectVariableWithID(projectID, key, value string, psd esv1.PushSecretData) error {
-	environmentScope := g.store.Environment
-	if environmentScope == "" {
-		environmentScope = "*"
-	}
-
-	// Extract metadata if provided
-	metadata := psd.GetMetadata()
-
-	// Check if variable exists
-	exists, err := g.projectVariableExistsWithID(projectID, key, environmentScope)
-	if err != nil {
-		return fmt.Errorf("failed to check if variable exists: %w", err)
-	}
-
-	if exists {
-		// Update existing variable
-		opts := &gitlab.UpdateProjectVariableOptions{
-			Value:            gitlab.Ptr(value),
-			EnvironmentScope: gitlab.Ptr(environmentScope),
-		}
-
-		// Apply metadata options if provided
-		applyMetadataToUpdateOptions(metadata, opts)
-
-		vopts := &gitlab.GetProjectVariableOptions{
-			Filter: &gitlab.VariableFilter{EnvironmentScope: environmentScope},
-		}
-
-		_, _, err := g.projectVariablesClient.UpdateVariable(projectID, key, opts, gitlab.WithContext(context.Background()))
-		metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabProjectVariableUpdate, err)
-		if err != nil {
-			// If update fails with 404, the variable might have a different environment scope
-			// Try to get and update with the correct scope
-			if errors.Is(err, gitlab.ErrNotFound) {
-				existingVar, _, getErr := g.projectVariablesClient.GetVariable(projectID, key, vopts)
-				if getErr == nil && existingVar != nil {
-					opts.Filter = &gitlab.VariableFilter{EnvironmentScope: existingVar.EnvironmentScope}
-					_, _, err = g.projectVariablesClient.UpdateVariable(projectID, key, opts)
-					metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabProjectVariableUpdate, err)
-				}
-			}
-		}
-		return err
-	}
-
-	// Create new variable
-	opts := &gitlab.CreateProjectVariableOptions{
-		Key:              gitlab.Ptr(key),
-		Value:            gitlab.Ptr(value),
-		EnvironmentScope: gitlab.Ptr(environmentScope),
-	}
-
-	// Apply metadata options if provided
-	applyMetadataToCreateOptions(metadata, opts)
-
-	_, _, err = g.projectVariablesClient.CreateVariable(projectID, opts)
-	metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabProjectVariableCreate, err)
-	return err
-}
-
-// pushProjectVariable creates or updates a project-level variable.
-func (g *gitlabBase) pushProjectVariable(key, value string, psd esv1.PushSecretData) error {
-	environmentScope := g.store.Environment
-	if environmentScope == "" {
-		environmentScope = "*"
-	}
-
-	// Extract metadata if provided
-	metadata := psd.GetMetadata()
-
-	// Check if variable exists
-	exists, err := g.projectVariableExists(key, environmentScope)
-	if err != nil {
-		return fmt.Errorf("failed to check if variable exists: %w", err)
-	}
-
-	if exists {
-		// Update existing variable
-		opts := &gitlab.UpdateProjectVariableOptions{
-			Value:            gitlab.Ptr(value),
-			EnvironmentScope: gitlab.Ptr(environmentScope),
-		}
-
-		// Apply metadata options if provided
-		applyMetadataToUpdateOptions(metadata, opts)
-
-		vopts := &gitlab.GetProjectVariableOptions{
-			Filter: &gitlab.VariableFilter{EnvironmentScope: environmentScope},
-		}
-
-		_, _, err := g.projectVariablesClient.UpdateVariable(g.store.ProjectID, key, opts, gitlab.WithContext(context.Background()))
-		metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabProjectVariableUpdate, err)
-		if err != nil {
-			// If update fails with 404, the variable might have a different environment scope
-			// Try to get and update with the correct scope
-			if errors.Is(err, gitlab.ErrNotFound) {
-				existingVar, _, getErr := g.projectVariablesClient.GetVariable(g.store.ProjectID, key, vopts)
-				if getErr == nil && existingVar != nil {
-					opts.Filter = &gitlab.VariableFilter{EnvironmentScope: existingVar.EnvironmentScope}
-					_, _, err = g.projectVariablesClient.UpdateVariable(g.store.ProjectID, key, opts)
-					metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabProjectVariableUpdate, err)
-				}
-			}
-		}
-		return err
-	}
-
-	// Create new variable
-	opts := &gitlab.CreateProjectVariableOptions{
-		Key:              gitlab.Ptr(key),
-		Value:            gitlab.Ptr(value),
-		EnvironmentScope: gitlab.Ptr(environmentScope),
-	}
-
-	// Apply metadata options if provided
-	applyMetadataToCreateOptions(metadata, opts)
-
-	_, _, err = g.projectVariablesClient.CreateVariable(g.store.ProjectID, opts)
-	metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabProjectVariableCreate, err)
-	return err
-}
-
-// pushGroupVariable creates or updates a group-level variable.
-func (g *gitlabBase) pushGroupVariable(groupID, key, value string, psd esv1.PushSecretData) error {
-	environmentScope := g.store.Environment
-	if environmentScope == "" {
-		environmentScope = "*"
-	}
-
-	// Extract metadata if provided
-	metadata := psd.GetMetadata()
-
-	// Check if variable exists
-	exists, err := g.groupVariableExists(groupID, key, environmentScope)
-	if err != nil {
-		return fmt.Errorf("failed to check if variable %q exists in group %q: %w", key, groupID, err)
-	}
-
-	if exists {
-		// Update existing variable
-		opts := &gitlab.UpdateGroupVariableOptions{
-			Value:            gitlab.Ptr(value),
-			EnvironmentScope: gitlab.Ptr(environmentScope),
-		}
-
-		// Apply metadata options if provided
-		applyMetadataToGroupUpdateOptions(metadata, opts)
-
-		_, _, err := g.groupVariablesClient.UpdateVariable(groupID, key, opts)
-		metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabGroupVariableUpdate, err)
-		if err != nil {
-			return fmt.Errorf("failed to update variable %q in group %q (env: %q): %w", key, groupID, environmentScope, err)
-		}
-		return nil
-	}
-
-	// Create new variable
-	opts := &gitlab.CreateGroupVariableOptions{
-		Key:              gitlab.Ptr(key),
-		Value:            gitlab.Ptr(value),
-		EnvironmentScope: gitlab.Ptr(environmentScope),
-	}
-
-	// Apply metadata options if provided
-	applyMetadataToGroupCreateOptions(metadata, opts)
-
-	_, _, err = g.groupVariablesClient.CreateVariable(groupID, opts)
-	metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabGroupVariableCreate, err)
-	if err != nil {
-		return fmt.Errorf("failed to create variable %q in group %q (env: %q): %w", key, groupID, environmentScope, err)
-	}
-	return nil
-}
-
-// projectVariableExists checks if a project variable exists.
-// projectVariableExists checks if a project variable exists in the default project.
-// Uses the project ID from the store configuration.
-func (g *gitlabBase) projectVariableExists(key, environmentScope string) (bool, error) {
-	return g.projectVariableExistsWithID(g.store.ProjectID, key, environmentScope)
-}
-
-// projectVariableExistsWithID checks if a project variable exists with a specific project ID.
-// Returns true if the variable exists, false if not found, or an error for other API failures.
-func (g *gitlabBase) projectVariableExistsWithID(projectID, key, environmentScope string) (bool, error) {
-	opts := &gitlab.GetProjectVariableOptions{
-		Filter: &gitlab.VariableFilter{EnvironmentScope: environmentScope},
-	}
-
-	_, resp, err := g.projectVariablesClient.GetVariable(projectID, key, opts)
+	// Retry with wildcard environment scope
+	opts := &gitlab.GetProjectVariableOptions{Filter: &gitlab.VariableFilter{EnvironmentScope: "*"}}
+	data, _, err = g.projectVariablesClient.GetVariable(g.store.ProjectID, ref.Key, opts)
 	metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabProjectVariableGet, err)
 
 	if err != nil {
-		if errors.Is(err, gitlab.ErrNotFound) || (resp != nil && resp.StatusCode == http.StatusNotFound) {
-			return false, nil
-		}
-		return false, err
+		return nil, fmt.Errorf("error getting variable %s from GitLab (including wildcard retry): %w", ref.Key, err)
 	}
 
-	return true, nil
+	return data, nil
 }
 
-// groupVariableExists checks if a group variable exists for the specified group.
-// Returns true if the variable exists, false if not found, or an error for other API failures.
-func (g *gitlabBase) groupVariableExists(groupID, key, environmentScope string) (bool, error) {
-	opts := &gitlab.GetGroupVariableOptions{
-		Filter: &gitlab.VariableFilter{EnvironmentScope: environmentScope},
-	}
-
-	_, resp, err := g.groupVariablesClient.GetVariable(groupID, key, opts)
-	metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabGroupGetVariable, err)
-
-	if err != nil {
-		if errors.Is(err, gitlab.ErrNotFound) || (resp != nil && resp.StatusCode == http.StatusNotFound) {
-			return false, nil
-		}
-		return false, err
-	}
-
-	return true, nil
-}
-
-// metadataOptions holds common GitLab variable options extracted from metadata.
-type metadataOptions struct {
-	Masked       *bool
-	Protected    *bool
-	Raw          *bool
-	Description  *string
-	VariableType *gitlab.VariableTypeValue
-}
-
-// parseMetadata extracts common GitLab variable options from PushSecret metadata.
-func parseMetadata(metadata *apiextensionsv1.JSON) *metadataOptions {
-	if metadata == nil {
-		return nil
-	}
-
-	var metadataMap map[string]interface{}
-	if err := json.Unmarshal(metadata.Raw, &metadataMap); err != nil {
-		return nil
-	}
-
-	opts := &metadataOptions{}
-
-	if masked, ok := metadataMap["masked"].(bool); ok {
-		opts.Masked = gitlab.Ptr(masked)
-	}
-	if protected, ok := metadataMap["protected"].(bool); ok {
-		opts.Protected = gitlab.Ptr(protected)
-	}
-	if raw, ok := metadataMap["raw"].(bool); ok {
-		opts.Raw = gitlab.Ptr(raw)
-	}
-	if description, ok := metadataMap["description"].(string); ok {
-		opts.Description = gitlab.Ptr(description)
-	}
-	if variableType, ok := metadataMap["variable_type"].(string); ok {
-		if variableType == "file" {
-			opts.VariableType = gitlab.Ptr(gitlab.FileVariableType)
-		} else {
-			opts.VariableType = gitlab.Ptr(gitlab.EnvVariableType)
-		}
-	}
-
-	return opts
-}
-
-// applyMetadataToCreateOptions applies metadata to project variable create options.
-func applyMetadataToCreateOptions(metadata *apiextensionsv1.JSON, opts *gitlab.CreateProjectVariableOptions) {
-	parsed := parseMetadata(metadata)
-	if parsed == nil {
-		return
-	}
-
-	if parsed.Masked != nil {
-		opts.Masked = parsed.Masked
-	}
-	if parsed.Protected != nil {
-		opts.Protected = parsed.Protected
-	}
-	if parsed.Raw != nil {
-		opts.Raw = parsed.Raw
-	}
-	if parsed.Description != nil {
-		opts.Description = parsed.Description
-	}
-	if parsed.VariableType != nil {
-		opts.VariableType = parsed.VariableType
-	}
-}
-
-// applyMetadataToUpdateOptions applies metadata to project variable update options.
-func applyMetadataToUpdateOptions(metadata *apiextensionsv1.JSON, opts *gitlab.UpdateProjectVariableOptions) {
-	parsed := parseMetadata(metadata)
-	if parsed == nil {
-		return
-	}
-
-	if parsed.Masked != nil {
-		opts.Masked = parsed.Masked
-	}
-	if parsed.Protected != nil {
-		opts.Protected = parsed.Protected
-	}
-	if parsed.Raw != nil {
-		opts.Raw = parsed.Raw
-	}
-	if parsed.Description != nil {
-		opts.Description = parsed.Description
-	}
-	if parsed.VariableType != nil {
-		opts.VariableType = parsed.VariableType
-	}
-}
-
-// applyMetadataToGroupCreateOptions applies metadata to group variable create options.
-func applyMetadataToGroupCreateOptions(metadata *apiextensionsv1.JSON, opts *gitlab.CreateGroupVariableOptions) {
-	parsed := parseMetadata(metadata)
-	if parsed == nil {
-		return
-	}
-
-	if parsed.Masked != nil {
-		opts.Masked = parsed.Masked
-	}
-	if parsed.Protected != nil {
-		opts.Protected = parsed.Protected
-	}
-	if parsed.Raw != nil {
-		opts.Raw = parsed.Raw
-	}
-	if parsed.Description != nil {
-		opts.Description = parsed.Description
-	}
-	if parsed.VariableType != nil {
-		opts.VariableType = parsed.VariableType
-	}
-}
-
-// applyMetadataToGroupUpdateOptions applies metadata to group variable update options.
-func applyMetadataToGroupUpdateOptions(metadata *apiextensionsv1.JSON, opts *gitlab.UpdateGroupVariableOptions) {
-	parsed := parseMetadata(metadata)
-	if parsed == nil {
-		return
-	}
-
-	if parsed.Masked != nil {
-		opts.Masked = parsed.Masked
-	}
-	if parsed.Protected != nil {
-		opts.Protected = parsed.Protected
-	}
-	if parsed.Raw != nil {
-		opts.Raw = parsed.Raw
-	}
-	if parsed.Description != nil {
-		opts.Description = parsed.Description
-	}
-	if parsed.VariableType != nil {
-		opts.VariableType = parsed.VariableType
-	}
-}
-
-// SecretExists checks if a secret exists in GitLab.
-// Note: PushSecretRemoteRef doesn't include metadata, so this uses store configuration only.
-func (g *gitlabBase) SecretExists(ctx context.Context, remoteRef esv1.PushSecretRemoteRef) (bool, error) {
-	if esutils.IsNil(g.projectVariablesClient) && esutils.IsNil(g.groupVariablesClient) {
-		return false, errors.New(errUninitializedGitlabProvider)
-	}
-
-	// Get the remote key and replace hyphens with underscores for GitLab API
-	remoteKey := strings.ReplaceAll(remoteRef.GetRemoteKey(), "-", "_")
-
-	environmentScope := g.store.Environment
-	if environmentScope == "" {
-		environmentScope = "*"
-	}
-
-	// Check project variable first if project ID is set
-	if g.store.ProjectID != "" {
-		exists, err := g.projectVariableExists(remoteKey, environmentScope)
-		if err != nil {
-			return false, err
-		}
-		if exists {
-			return true, nil
-		}
-	}
-
-	// Check group variables if configured
-	if len(g.store.GroupIDs) > 0 {
-		exists, err := g.groupVariableExists(g.store.GroupIDs[0], remoteKey, environmentScope)
-		if err != nil {
-			return false, err
-		}
-		return exists, nil
-	}
-
-	return false, nil
-}
-
-// DeleteSecret deletes a variable from GitLab.
-// LIMITATION: This can only delete from the store's configured projectID or first groupID.
-// Variables pushed with metadata-based groupID/projectID overrides cannot be automatically
-// deleted because the SecretsClient.DeleteSecret interface doesn't provide metadata.
-// Users must either:
-// 1. Not use deletionPolicy: Delete with metadata overrides, OR
-// 2. Manually clean up variables pushed to non-default locations
-func (g *gitlabBase) DeleteSecret(ctx context.Context, remoteRef esv1.PushSecretRemoteRef) error {
-	if esutils.IsNil(g.projectVariablesClient) && esutils.IsNil(g.groupVariablesClient) {
-		return errors.New(errUninitializedGitlabProvider)
-	}
-
-	// Get the remote key and replace hyphens with underscores for GitLab API
-	remoteKey := strings.ReplaceAll(remoteRef.GetRemoteKey(), "-", "_")
-
-	environmentScope := g.store.Environment
-	if environmentScope == "" {
-		environmentScope = "*"
-	}
-
-	// Only delete from the store's default location (no metadata available)
-	if g.store.ProjectID != "" {
-		opts := &gitlab.RemoveProjectVariableOptions{
-			Filter: &gitlab.VariableFilter{EnvironmentScope: environmentScope},
-		}
-		_, err := g.projectVariablesClient.RemoveVariable(g.store.ProjectID, remoteKey, opts)
-		metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabProjectVariableDelete, err)
-
-		if err == nil || errors.Is(err, gitlab.ErrNotFound) {
-			return nil
-		}
-		return fmt.Errorf("failed to delete from project %q: %w", g.store.ProjectID, err)
-	}
-
-	// Delete from first configured group only
-	if len(g.store.GroupIDs) > 0 {
-		opts := &gitlab.RemoveGroupVariableOptions{
-			Filter: &gitlab.VariableFilter{EnvironmentScope: environmentScope},
-		}
-		_, err := g.groupVariablesClient.RemoveVariable(g.store.GroupIDs[0], remoteKey, opts)
-		metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabGroupVariableDelete, err)
-
-		if err == nil || errors.Is(err, gitlab.ErrNotFound) {
-			return nil
-		}
-		return fmt.Errorf("failed to delete from group %q: %w", g.store.GroupIDs[0], err)
-	}
-
-	return fmt.Errorf("%s for delete operation", errNoProjectOrGroup)
-}
-
-// GetAllSecrets syncs all gitlab project and group variables into a single Kubernetes Secret.
-func (g *gitlabBase) GetAllSecrets(_ context.Context, ref esv1.ExternalSecretFind) (map[string][]byte, error) {
-	if esutils.IsNil(g.projectVariablesClient) {
-		return nil, errors.New(errUninitializedGitlabProvider)
-	}
-	var effectiveEnvironment = g.store.Environment
-	if ref.Tags != nil {
-		environment, err := ExtractTag(ref.Tags)
-		if err != nil {
-			return nil, err
-		}
-		if !isEmptyOrWildcard(effectiveEnvironment) && !isEmptyOrWildcard(environment) {
-			return nil, errors.New(errEnvironmentIsConstricted)
-		}
-		effectiveEnvironment = environment
-	}
-	if ref.Path != nil {
-		return nil, errors.New(errPathNotImplemented)
-	}
-	if ref.Name == nil {
-		return nil, errors.New(errNameNotDefined)
-	}
-
-	var matcher *find.Matcher
-	if ref.Name != nil {
-		m, err := find.New(*ref.Name)
-		if err != nil {
-			return nil, err
-		}
-		matcher = m
-	}
-
-	err := g.ResolveGroupIDs()
-	if err != nil {
-		return nil, err
-	}
-
-	secretData, err := g.fetchSecretData(effectiveEnvironment, matcher)
-	if err != nil {
-		return nil, err
-	}
-
-	// _Note_: fetchProjectVariables alters secret data map
-	if err := g.fetchProjectVariables(effectiveEnvironment, matcher, secretData); err != nil {
-		return nil, err
-	}
-
-	return secretData, nil
-}
-
-func (g *gitlabBase) fetchProjectVariables(effectiveEnvironment string, matcher *find.Matcher, secretData map[string][]byte) error {
-	var popts = &gitlab.ListProjectVariablesOptions{PerPage: 100}
-	nonWildcardSet := make(map[string]bool)
-	for projectPage := 1; ; projectPage++ {
-		popts.Page = projectPage
-		projectData, response, err := g.projectVariablesClient.ListVariables(g.store.ProjectID, popts)
-		metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabProjectListVariables, err)
-		if err != nil {
-			return err
-		}
-
-		processProjectVariables(projectData, effectiveEnvironment, matcher, secretData, nonWildcardSet)
-		if response.CurrentPage >= response.TotalPages {
-			break
-		}
-	}
-
-	return nil
-}
-
-func processProjectVariables(
-	projectData []*gitlab.ProjectVariable,
-	effectiveEnvironment string,
-	matcher *find.Matcher,
-	secretData map[string][]byte,
-	nonWildcardSet map[string]bool,
-) {
-	for _, data := range projectData {
-		matching, key, isWildcard := matchesFilter(effectiveEnvironment, data.EnvironmentScope, data.Key, matcher)
-		if !matching {
-			continue
-		}
-		if isWildcard && nonWildcardSet[key] {
-			continue
-		}
-		secretData[key] = []byte(data.Value)
-		if !isWildcard {
-			nonWildcardSet[key] = true
-		}
-	}
-}
-
-func (g *gitlabBase) fetchSecretData(effectiveEnvironment string, matcher *find.Matcher) (map[string][]byte, error) {
-	var gopts = &gitlab.ListGroupVariablesOptions{PerPage: 100}
-	secretData := make(map[string][]byte)
-	for _, groupID := range g.store.GroupIDs {
-		if err := g.setVariablesForGroupID(effectiveEnvironment, matcher, gopts, groupID, secretData); err != nil {
-			return nil, err
-		}
-	}
-
-	return secretData, nil
-}
-
-func (g *gitlabBase) setVariablesForGroupID(
-	effectiveEnvironment string,
-	matcher *find.Matcher,
-	gopts *gitlab.ListGroupVariablesOptions,
-	groupID string,
-	secretData map[string][]byte,
-) error {
-	for groupPage := 1; ; groupPage++ {
-		gopts.Page = groupPage
-		groupVars, response, err := g.groupVariablesClient.ListVariables(groupID, gopts)
-		metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabGroupListVariables, err)
-		if err != nil {
-			return err
-		}
-		g.setGroupValues(effectiveEnvironment, matcher, groupVars, secretData)
-
-		if response.CurrentPage >= response.TotalPages {
-			break
-		}
-	}
-	return nil
-}
-
-func (g *gitlabBase) setGroupValues(
-	effectiveEnvironment string,
-	matcher *find.Matcher,
-	groupVars []*gitlab.GroupVariable,
-	secretData map[string][]byte,
-) {
-	for _, data := range groupVars {
-		matching, key, isWildcard := matchesFilter(effectiveEnvironment, data.EnvironmentScope, data.Key, matcher)
-		if !matching {
-			continue
-		}
-		// Check if a more specific variable already exists (project environment > project variable > group environment > group variable)
-		_, exists := secretData[key]
-		if exists && isWildcard {
-			continue
-		}
-		secretData[key] = []byte(data.Value)
-	}
-}
-
-// ExtractTag extracts the environment scope from the provided tags map.
-func ExtractTag(tags map[string]string) (string, error) {
-	var environmentScope string
-	for tag, value := range tags {
-		if tag != "environment_scope" {
-			return "", errors.New(errTagsOnlyEnvironmentSupported)
-		}
-		environmentScope = value
-	}
-	return environmentScope, nil
-}
-
-func (g *gitlabBase) getGroupVariables(groupID string, ref esv1.ExternalSecretDataRemoteRef, gopts *gitlab.GetGroupVariableOptions) (*gitlab.GroupVariable, *gitlab.Response, error) {
-	groupVar, resp, err := g.groupVariablesClient.GetVariable(groupID, ref.Key, gopts)
-	metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabGroupGetVariable, err)
-	if err != nil {
-		if resp != nil && resp.StatusCode == http.StatusNotFound && !isEmptyOrWildcard(g.store.Environment) {
-			if gopts == nil {
-				gopts = &gitlab.GetGroupVariableOptions{}
-			}
-			if gopts.Filter == nil {
-				gopts.Filter = &gitlab.VariableFilter{}
-			}
-			gopts.Filter.EnvironmentScope = "*"
-			groupVar, resp, err = g.groupVariablesClient.GetVariable(groupID, ref.Key, gopts)
-			metrics.ObserveAPICall(constants.ProviderGitLab, constants.CallGitLabGroupGetVariable, err)
-			if err != nil || resp == nil {
-				return nil, resp, fmt.Errorf("error getting group variable %s from GitLab: %w", ref.Key, err)
-			}
-		} else {
-			return nil, resp, err
-		}
-	}
-
-	return groupVar, resp, nil
-}
-
-func (g *gitlabBase) GetSecret(_ context.Context, ref esv1.ExternalSecretDataRemoteRef) ([]byte, error) {
-	if esutils.IsNil(g.projectVariablesClient) || esutils.IsNil(g.groupVariablesClient) {
-		return nil, errors.New(errUninitializedGitlabProvider)
-	}
-
-	// Need to replace hyphens with underscores to work with GitLab API
-	ref.Key = strings.ReplaceAll(ref.Key, "-", "_")
-	// Retrieves a gitlab variable in the form
-	// {
-	// 	"key": "TEST_VARIABLE_1",
-	// 	"variable_type": "env_var",
-	// 	"value": "TEST_1",
-	// 	"protected": false,
-	// 	"masked": true,
-	// 	"environment_scope": "*"
-	// }
-	var gopts *gitlab.GetGroupVariableOptions
-	var vopts *gitlab.GetProjectVariableOptions
-	if g.store.Environment != "" {
-		gopts = &gitlab.GetGroupVariableOptions{Filter: &gitlab.VariableFilter{EnvironmentScope: g.store.Environment}}
-		vopts = &gitlab.GetProjectVariableOptions{Filter: &gitlab.VariableFilter{EnvironmentScope: g.store.Environment}}
-	}
-
-	data, err := g.getVariables(ref, vopts)
-	if err == nil {
-		return extractVariable(ref, data.Value)
-	}
-
-	// If project variable not found, try group variables
-	if errors.Is(err, gitlab.ErrNotFound) {
-		return g.tryGroupVariables(ref, gopts, err)
-	}
-
-	return nil, err
-}
-
-// tryGroupVariables attempts to retrieve the secret from group variables when project lookup fails.
-func (g *gitlabBase) tryGroupVariables(ref esv1.ExternalSecretDataRemoteRef, gopts *gitlab.GetGroupVariableOptions, originalErr error) ([]byte, error) {
-	// Load groupIds from the `InheritFromGroups` property
-	if err := g.ResolveGroupIDs(); err != nil {
-		return nil, err
-	}
-
-	for i := len(g.store.GroupIDs) - 1; i >= 0; i-- {
-		groupID := g.store.GroupIDs[i]
-		groupVar, _, err := g.getGroupVariables(groupID, ref, gopts)
-		if err == nil {
-			return extractVariable(ref, groupVar.Value)
-		}
-
-		// If a 404 error, continue to the next stage, otherwise exit early with error
-		if errors.Is(err, gitlab.ErrNotFound) {
-			continue
-		}
-		return nil, err
-	}
-
-	// No group variables found, return the original project error
-	return nil, originalErr
-}
-
-func extractVariable(ref esv1.ExternalSecretDataRemoteRef, value string) ([]byte, error) {
-	// If no property specified, return the raw value
-	if ref.Property == "" {
-		if value == "" {
-			return nil, fmt.Errorf("invalid secret received. no secret string for key: %s", ref.Key)
-		}
-		return []byte(value), nil
-	}
-
-	// Extract property from JSON value
-	val := gjson.Get(value, ref.Property)
-	if !val.Exists() {
-		return nil, fmt.Errorf("key %s does not exist in secret %s", ref.Property, ref.Key)
-	}
-	return []byte(val.String()), nil
-}
-
-func (g *gitlabBase) GetSecretMap(ctx context.Context, ref esv1.ExternalSecretDataRemoteRef) (map[string][]byte, error) {
-	// Gets a secret as normal, expecting secret value to be a json object
-	data, err := g.GetSecret(ctx, ref)
-	if err != nil {
-		return nil, fmt.Errorf("error getting secret %s: %w", ref.Key, err)
-	}
-
-	// Maps the json data to a string:string map
-	kv := make(map[string]string)
-	err = json.Unmarshal(data, &kv)
-	if err != nil {
-		return nil, fmt.Errorf(errJSONSecretUnmarshal, err)
-	}
-
-	// Converts values in K:V pairs into bytes, while leaving keys as strings
-	secretData := make(map[string][]byte)
-	for k, v := range kv {
-		secretData[k] = []byte(v)
-	}
-	return secretData, nil
-}
-
-func isEmptyOrWildcard(environment string) bool {
-	return environment == "" || environment == "*"
-}
-
-func matchesFilter(environment, varEnvironment, key string, matcher *find.Matcher) (bool, string, bool) {
-	isWildcard := isEmptyOrWildcard(varEnvironment)
-	if !isWildcard && !isEmptyOrWildcard(environment) {
-		// as of now gitlab does not support filtering of EnvironmentScope through the api call
-		if varEnvironment != environment {
-			return false, "", isWildcard
-		}
-	}
-
-	if key == "" || (matcher != nil && !matcher.MatchName(key)) {
-		return false, "", isWildcard
-	}
-
-	return true, key, isWildcard
-}
-
+// Close closes the client (no-op for GitLab).
 func (g *gitlabBase) Close(_ context.Context) error {
 	return nil
 }
 
+// ResolveGroupIDs resolves group IDs when inheritFromGroups is enabled.
 func (g *gitlabBase) ResolveGroupIDs() error {
 	if g.store.InheritFromGroups {
 		projectGroups, resp, err := g.projectsClient.ListProjectsGroups(g.store.ProjectID, nil)
@@ -950,7 +145,7 @@ func (g *gitlabBase) ResolveGroupIDs() error {
 	return nil
 }
 
-// Validate will use the gitlab projectVariablesClient/groupVariablesClient to validate the gitlab provider using the ListVariable call to ensure get permissions without needing a specific key.
+// Validate validates the gitlab provider credentials and permissions.
 func (g *gitlabBase) Validate() (esv1.ValidationResult, error) {
 	if g.store.ProjectID != "" {
 		_, resp, err := g.projectVariablesClient.ListVariables(g.store.ProjectID, nil)
